Fix for wrong variable type change

+++ src/lib/sup-mvvm-viewmodel/mvvm/editors/scientific_spinbox.h
@@ -72,8 +72,8 @@ private:
   bool inRange(double val) const;
 
   double m_value;
-  int m_min;
-  int m_max;
+  double m_min;
+  double m_max;
   double m_step;
   int m_decimals;
   QDoubleValidator m_validator;
